Fix conflicting recipes.
In testing I noticed that you could have a log with an axe *and* a saw, and this isn't the intended behavior.

This commit does come with the drawback of up to 7 iterations in a for-loop that can no longer be skipped - this should be a neglegable performance decrease. This commit makes it so that the recipe is invalid if both an axe and saw are included in the grid with a log.