Remove recipes the *right* way
Upon reaching out on the MinecraftForge forums (https://bit.ly/2Ns6Vkv), I learned that the method I was using before by catching if a log was in the workbench, claiming the recipe belonged to my custom handler, then setting an empty output was *not* the solution, causing inconsistency with recipe outputs sometimes not working the way I intended.

Choonster (https://github.com/Choonster-Minecraft-Mods) provided the solution, so this commit includes a custom implementation of his solution with itemstack-specific recipe removal (for wooden tools). This will allow me to make recipe disabling configurable in the future as well.